refactor: Use docker compose for insert tests
