+++ testdata/docker/spark-script/spark-connect-server.sh
@@ -18,7 +18,7 @@ PACKAGES="$PACKAGES,org.apache.spark:spark-connect_2.12:$SPARK_VERSION"
   --conf spark.sql.catalog.demo=org.apache.iceberg.spark.SparkCatalog \
   --conf spark.sql.catalog.demo.type=hadoop \
   --conf spark.sql.catalog.demo.warehouse=s3a://icebergdata/demo \
-  --conf spark.sql.catalog.demo.hadoop.fs.s3a.endpoint=http://"$MINIO_IP":9000 \
+  --conf spark.sql.catalog.demo.hadoop.fs.s3a.endpoint=http://minio:9000 \
   --conf spark.sql.catalog.demo.hadoop.fs.s3a.path.style.access=true \
   --conf spark.sql.catalog.demo.hadoop.fs.s3a.access.key=admin \
   --conf spark.sql.catalog.demo.hadoop.fs.s3a.secret.key=password \
